# frontend/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConfiguracionLocalDto": {
        "properties": {
          "calificacionesParaCupon": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "diasVigenciaCupon": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "minutosObjetivoCocina": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "porcentajeDescuentoCupon": {
            "exclusiveMaximum": false,
            "exclusiveMinimum": false,
            "maximum": 100,
            "minimum": 0,
            "type": "number"
          }
        },
        "type": "object"
      },
...
      "CuponResponseDto": {
        "properties": {
          "clienteId": {
            "format": "uuid",
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "estado": {
            "enum": [
              "VIGENTE",
              "CANJEADO",
              "VENCIDO",
              "ANULADO"
            ],
            "type": "string"
          },
          "fechaEmision": {
            "format": "date-time",
            "type": "string"
          },
          "fechaVencimiento": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "montoDescuento": {
            "type": "number"
          },
          "porcentajeDescuento": {
            "type": "number"
          },
          "vigente": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "FeedbackRequestDto": {
        "properties": {
          "comentario": {
            "type": "string"
          },
          "puntajeAtencion": {
            "format": "int32",
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "puntajeComida": {
            "format": "int32",
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "puntajeLugar": {
            "format": "int32",
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "puntajeAtencion",
          "puntajeComida",
          "puntajeLugar"
        ],
        "type": "object"
      },
      "FeedbackResponseDto": {
        "properties": {
          "clienteId": {
            "format": "uuid",
            "type": "string"
          },
          "comentario": {
            "type": "string"
          },
          "cuponGenerado": {
            "$ref": "#/components/schemas/CuponResponseDto"
          },
          "fecha": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "mensajeFidelizacion": {
            "type": "string"
          },
          "ordenId": {
            "format": "uuid",
            "type": "string"
          },
          "promedio": {
            "format": "double",
            "type": "number"
          },
          "puntajeAtencion": {
            "format": "int32",
            "type": "integer"
          },
          "puntajeComida": {
            "format": "int32",
            "type": "integer"
          },
          "puntajeLugar": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "FidelizacionDto": {
        "properties": {
          "calificacionesFaltantes": {
            "format": "int64",
            "type": "integer"
          },
          "calificacionesRealizadas": {
            "format": "int64",
            "type": "integer"
          },
          "calificacionesRequeridas": {
            "format": "int32",
            "type": "integer"
          },
          "clienteId": {
            "format": "uuid",
            "type": "string"
          },
          "cuponesVigentes": {
            "format": "int64",
            "type": "integer"
          },
          "mensaje": {
            "type": "string"
          },
          "puntosFidelidad": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "OrdenResumenDto": {
        "properties": {
          "canalOrigen": {
            "enum": [
              "POS",
              "WHATSAPP_BOT",
              "DISCORD_BOT",
              "WEB"
            ],
            "type": "string"
          },
          "cantidadItems": {
            "format": "int32",
            "type": "integer"
          },
          "clienteNombre": {
            "type": "string"
          },
          "editable": {
            "type": "boolean"
          },
          "estado": {
            "enum": [
              "ENCOLADO",
              "EN_PREPARACION",
              "EN_DESPACHO",
              "ENTREGADO",
              "PAGADO",
              "CONCLUIDO",
              "CANCELADO",
              "FRAUDULENTO"
            ],
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "mesaId": {
            "format": "uuid",
            "type": "string"
          },
          "mesaNumero": {
            "type": "string"
          },
          "minutosTranscurridos": {
            "format": "int64",
            "type": "integer"
          },
          "montoTotal": {
            "type": "number"
          },
          "tiempoInicioGlobal": {
            "format": "date-time",
            "type": "string"
          },
          "tipoOrden": {
            "enum": [
              "MESA",
              "RETIRO_LOCAL",
              "DELIVERY"
            ],
            "type": "string"
          },
          "tipoPago": {
            "enum": [
              "EFECTIVO",
              "E_WALLET",
              "TARJETA"
            ],
            "type": "string"
          },
          "transicionesPermitidas": {
            "items": {
              "enum": [
                "ENCOLADO",
                "EN_PREPARACION",
                "EN_DESPACHO",
                "ENTREGADO",
                "PAGADO",
                "CONCLUIDO",
                "CANCELADO",
                "FRAUDULENTO"
              ],
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "PageResponseDtoTransportistaResponseDto": {
        "properties": {
          "contenido": {
            "items": {
              "$ref": "#/components/schemas/TransportistaResponseDto"
            },
            "type": "array"
          },
          "pagina": {
            "format": "int32",
            "type": "integer"
          },
          "primera": {
            "type": "boolean"
          },
          "tamano": {
            "format": "int32",
            "type": "integer"
          },
          "totalElementos": {
            "format": "int64",
            "type": "integer"
          },
          "totalPaginas": {
            "format": "int32",
            "type": "integer"
          },
          "ultima": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Pageable": {
        "properties": {
          "page": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "minimum": 1,
            "type": "integer"
          },
          "sort": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "ReporteSatisfaccionDto": {
        "properties": {
          "desde": {
            "format": "date-time",
            "type": "string"
          },
          "hasta": {
            "format": "date-time",
            "type": "string"
          },
          "promedioAtencion": {
            "format": "double",
            "type": "number"
          },
          "promedioComida": {
            "format": "double",
            "type": "number"
          },
          "promedioGeneral": {
            "format": "double",
            "type": "number"
          },
          "promedioLugar": {
            "format": "double",
            "type": "number"
          },
          "totalCalificaciones": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "TransportistaRequestDto": {
        "properties": {
          "apellidos": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "celular": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "correo": {
            "format": "email",
            "type": "string"
          },
          "dni": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "empresaTransporte": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "nombres": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "apellidos",
          "celular",
          "dni",
          "empresaTransporte",
          "nombres"
        ],
        "type": "object"
      },
      "TransportistaResponseDto": {
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "apellidos": {
            "type": "string"
          },
          "celular": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "dni": {
            "type": "string"
          },
          "empresaTransporte": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nombreCompleto": {
            "type": "string"
          },
          "nombres": {
            "type": "string"
          },
          "vehiculos": {
            "items": {
              "$ref": "#/components/schemas/VehiculoResponseDto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "VehiculoRequestDto": {
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "marcaModelo": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "placa": {
            "maxLength": 15,
            "minLength": 0,
            "type": "string"
          },
          "tipoVehiculo": {
            "enum": [
              "MOTO",
              "TAXI_AUTO",
              "BICICLETA"
            ],
            "type": "string"
          }
        },
        "required": [
          "placa",
          "tipoVehiculo"
        ],
        "type": "object"
      },
      "VehiculoResponseDto": {
        "properties": {
          "activo": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "marcaModelo": {
            "type": "string"
          },
          "placa": {
            "type": "string"
          },
          "tipoVehiculo": {
            "enum": [
              "MOTO",
              "TAXI_AUTO",
              "BICICLETA"
            ],
            "type": "string"
          },
          "transportistaId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API de POS, inventario, comandas, cocina, despacho y caja del sistema de gestion gastronomica Chaquena.",
    "title": "Backend Logistica - Chaquena",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/clientes/{clienteId}/cupones": {
      "get": {
        "operationId": "cupones",
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CuponResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Cupones del cliente",
        "tags": [
          "Feedback y fidelizacion"
        ]
      }
    },
    "/api/v1/clientes/{clienteId}/fidelizacion": {
      "get": {
        "operationId": "progreso",
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FidelizacionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Cuantas calificaciones lleva el cliente y cuantas le faltan",
        "tags": [
          "Feedback y fidelizacion"
        ]
      }
    },
...
    "/api/v1/clientes/{id}/ordenes": {
      "get": {
        "operationId": "ordenes",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrdenResumenDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Historial de comandas del cliente",
        "tags": [
          "Clientes"
        ]
      }
    },
...
    "/api/v1/configuracion": {
      "get": {
        "operationId": "configuracion",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionLocalDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Parametros del local: umbral N, descuento y objetivo de cocina",
        "tags": [
          "Feedback y fidelizacion"
        ]
      },
      "put": {
        "operationId": "actualizarConfiguracion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfiguracionLocalDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracionLocalDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Actualizar los parametros del local",
        "tags": [
          "Feedback y fidelizacion"
        ]
      }
    },
    "/api/v1/cupones/{codigo}/canjear": {
      "post": {
        "operationId": "canjear",
        "parameters": [
          {
            "in": "path",
            "name": "codigo",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ordenId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuponResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Canjear cupon contra una comanda",
        "tags": [
          "Feedback y fidelizacion"
        ]
      }
    },
...
    "/api/v1/ordenes/{ordenId}/feedback": {
      "get": {
        "operationId": "obtenerFeedbackDeOrden",
        "parameters": [
          {
            "in": "path",
            "name": "ordenId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Calificacion de una comanda",
        "tags": [
          "Feedback y fidelizacion"
        ]
      },
      "post": {
        "operationId": "registrarFeedbackDeOrden",
        "parameters": [
          {
            "in": "path",
            "name": "ordenId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Registrar calificacion y evaluar la regla de las N calificaciones",
        "tags": [
          "Feedback y fidelizacion"
        ]
      }
    },
...
    "/api/v1/reportes/satisfaccion": {
      "get": {
        "operationId": "satisfaccion",
        "parameters": [
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReporteSatisfaccionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Promedios de satisfaccion por criterio y periodo",
        "tags": [
          "Feedback y fidelizacion"
        ]
      }
    },
...
    "/api/v1/transportistas": {
      "get": {
        "operationId": "listarTransportistas",
        "parameters": [
          {
            "in": "query",
            "name": "empresa",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "pageable",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDtoTransportistaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Listar transportistas",
        "tags": [
          "Despacho - Transportistas"
        ]
      },
      "post": {
        "operationId": "crearTransportista",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportistaRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportistaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Registrar transportista con DNI, empresa y telefono",
        "tags": [
          "Despacho - Transportistas"
        ]
      }
    },
...
    "/api/v1/transportistas/{id}": {
      "put": {
        "operationId": "actualizarTransportista",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportistaRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportistaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Actualizar datos del transportista",
        "tags": [
          "Despacho - Transportistas"
        ]
      }
    },
    "/api/v1/transportistas/{id}/activo": {
      "patch": {
        "operationId": "cambiarActivoTransportista",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "activo",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportistaResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Habilitar o dar de baja al transportista",
        "tags": [
          "Despacho - Transportistas"
        ]
      }
    },
    "/api/v1/transportistas/{id}/vehiculos": {
      "get": {
        "operationId": "vehiculos",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VehiculoResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Vehiculos del transportista",
        "tags": [
          "Despacho - Transportistas"
        ]
      },
      "post": {
        "operationId": "registrarVehiculo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoResponseDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Registrar placa, tipo y modelo del vehiculo",
        "tags": [
          "Despacho - Transportistas"
        ]
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}